+++ src/moab/Core.hpp
@@ -1224,16 +1224,16 @@ public:
    */
   void estimated_memory_use( const EntityHandle* ent_array = 0,
                              unsigned long  num_ents = 0,
-                             unsigned long* total_storage = 0,
-                             unsigned long* total_amortized_storage = 0,
-                             unsigned long* entity_storage = 0,
-                             unsigned long* amortized_entity_storage = 0,
-                             unsigned long* adjacency_storage = 0,
-                             unsigned long* amortized_adjacency_storage = 0,
+                             unsigned long long* total_storage = 0,
+                             unsigned long long* total_amortized_storage = 0,
+                             unsigned long long* entity_storage = 0,
+                             unsigned long long* amortized_entity_storage = 0,
+                             unsigned long long* adjacency_storage = 0,
+                             unsigned long long* amortized_adjacency_storage = 0,
                              const Tag*   tag_array = 0,
                              unsigned       num_tags = 0,
-                             unsigned long* tag_storage = 0,
-                             unsigned long* amortized_tag_storage = 0 );
+                             unsigned long long* tag_storage = 0,
+                             unsigned long long* amortized_tag_storage = 0 );
 
   /**\brief Calculate amount of memory used to store MOAB data
    *
@@ -1262,16 +1262,16 @@ public:
    *                   all tags.
    */
   void estimated_memory_use( const Range& ents,
-                             unsigned long* total_storage = 0,
-                             unsigned long* total_amortized_storage = 0,
-                             unsigned long* entity_storage = 0,
-                             unsigned long* amortized_entity_storage = 0,
-                             unsigned long* adjacency_storage = 0,
-                             unsigned long* amortized_adjacency_storage = 0,
+                             unsigned long long* total_storage = 0,
+                             unsigned long long* total_amortized_storage = 0,
+                             unsigned long long* entity_storage = 0,
+                             unsigned long long* amortized_entity_storage = 0,
+                             unsigned long long* adjacency_storage = 0,
+                             unsigned long long* amortized_adjacency_storage = 0,
                              const Tag*   tag_array = 0,
                              unsigned       num_tags = 0,
-                             unsigned long* tag_storage = 0,
-                             unsigned long* amortized_tag_storage = 0 );
+                             unsigned long long* tag_storage = 0,
+                             unsigned long long* amortized_tag_storage = 0 );
                                      
 
   void print_database() const;
@@ -1284,16 +1284,16 @@ private:
   Core& operator=( const Core& copy );
 
   void estimated_memory_use_internal( const Range* ents,
-                            unsigned long* total_storage,
-                            unsigned long* total_amortized_storage,
-                            unsigned long* entity_storage,
-                            unsigned long* amortized_entity_storage,
-                            unsigned long* adjacency_storage,
-                            unsigned long* amortized_adjacency_storage,
+                            unsigned long long* total_storage,
+                            unsigned long long* total_amortized_storage,
+                            unsigned long long* entity_storage,
+                            unsigned long long* amortized_entity_storage,
+                            unsigned long long* adjacency_storage,
+                            unsigned long long* amortized_adjacency_storage,
                             const Tag*   tag_array,
                             unsigned       num_tags,
-                            unsigned long* tag_storage,
-                            unsigned long* amortized_tag_storage );
+                            unsigned long long* tag_storage,
+                            unsigned long long* amortized_tag_storage );
 
     //! database init and de-init routines
   ErrorCode initialize();
